requirements.txt
azure-ai-inference
azure-core
scikit-learn
fpdf2
python-dotenv
pdfplumber
python-docx
PyMuPDF